Create database test module and move BaseDataTest classes there.
