+++ database/test/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+  alias(libs.plugins.tvmaniac.kotlin.android)
+  alias(libs.plugins.tvmaniac.multiplatform)
+  alias(libs.plugins.sqldelight)
+}
+
+kotlin {
+  sourceSets {
+    androidMain { dependencies { implementation(libs.sqldelight.driver.jvm) } }
+
+    commonMain {
+      dependencies {
+        implementation(projects.database)
+
+        implementation(libs.kotlin.test)
+      }
+    }
+
+    iosMain { dependencies { implementation(libs.sqldelight.driver.native) } }
+
+    jvmMain { dependencies { implementation(libs.sqldelight.driver.jvm) } }
+  }
+}
+
+android { namespace = "com.thomaskioko.tvmaniac.database.test" }
