+++ database/test/src/iosMain/kotlin/com/thomaskioko/tvmaniac/database/test/BaseDatabaseTest.kt
@@ -1,9 +1,10 @@
-package com.thomaskioko.tvmaniac.core.db
+package com.thomaskioko.tvmaniac.database.test
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.native.NativeSqliteDriver
 import app.cash.sqldelight.driver.native.wrapConnection
 import co.touchlab.sqliter.DatabaseConfiguration
+import com.thomaskioko.tvmaniac.core.db.TvManiacDatabase
 
 actual fun inMemorySqlDriver(): SqlDriver =
   NativeSqliteDriver(
